Extract Retirable mixin
