+++ app/controllers/api/services_controller.rb
@@ -2,6 +2,7 @@ module Api
   class ServicesController < BaseController
     include Shared::CustomActions
     include Shared::Ownable
+    include Shared::Retirable
     include Subcollections::ServiceDialogs
     include Subcollections::Tags
     include Subcollections::Vms
