+++ app/controllers/api/vms_controller.rb
@@ -1,6 +1,7 @@
 module Api
   class VmsController < BaseController
     include Shared::Ownable
+    include Shared::Retirable
     include Subcollections::Tags
     include Subcollections::Policies
     include Subcollections::PolicyProfiles
